refactor: Move stream/definition/column to service/stream/mapping/table/column
