+++ internal/pkg/service/stream/mapping/table/column/body.go
