+++ internal/pkg/service/stream/mapping/table/column/column.go
