+++ internal/pkg/service/stream/mapping/table/column/columns.go
