+++ internal/pkg/service/stream/mapping/table/column/headers.go
