+++ internal/pkg/service/stream/mapping/table/column/template.go
